First
An online perfume sales site that uses springboot and postgresql on the backend and html css and javascript on the front. As it is a trial project, the site does not have real purchasing functions, but the frontend is still interactive. If you want to see the site with its products, I also share the data with you.